fix: expose explorer drag data
