+++ packages/explorer-view/src/parts/CommandMap/CommandMap.ts
@@ -19,6 +19,7 @@ import * as FocusLast from '../FocusLast/FocusLast.ts'
 import * as FocusNext from '../FocusNext/FocusNext.ts'
 import * as FocusNone from '../FocusNone/FocusNone.ts'
 import * as FocusPrevious from '../FocusPrevious/FocusPrevious.ts'
+import { getCurrentDragData } from '../GetCurrentDragData/GetCurrentDragData.ts'
 import * as GetKeyBindings from '../GetKeyBindings/GetKeyBindings.ts'
 import * as GetMenuEntries2 from '../GetMenuEntries2/GetMenuEntries2.ts'
 import * as GetMenuEntries from '../GetMenuEntries/GetMenuEntries.ts'
@@ -112,6 +113,7 @@ export const commandMap = {
   'Explorer.focusNone': WrapCommand.wrapListItemCommand(FocusNone.focusNone),
   'Explorer.focusPrevious': WrapCommand.wrapListItemCommand(FocusPrevious.focusPrevious),
   'Explorer.getCommandIds': WrapCommand.getCommandIds,
+  'Explorer.getDragData': WrapCommand.wrapGetter(getCurrentDragData),
   'Explorer.getKeyBindings': GetKeyBindings.getKeyBindings,
   'Explorer.getMenuEntries': GetMenuEntries.getMenuEntries,
   'Explorer.getMenuEntries2': WrapCommand.wrapGetter(GetMenuEntries2.getMenuEntries2),
